CLN: move predict_transport_mode
