+++ trackintel/analysis/__init__.py
@@ -1,3 +1,2 @@
-from .transport_mode_identification import predict_transport_mode
-from .tracking_quality import temporal_tracking_quality
-from .label import create_activity_flag
+from .tracking_quality import *
+from .label import *
