+++ trackintel/model/triplegs.py
@@ -107,9 +107,9 @@ class TriplegsAccessor(object):
         """
         Predict/impute the transport mode with which each tripleg was likely covered.
 
-        See :func:`trackintel.analysis.transport_mode_identification.predict_transport_mode`.
+        See :func:`trackintel.analysis.label.predict_transport_mode`.
         """
-        return ti.analysis.transport_mode_identification.predict_transport_mode(self._obj, *args, **kwargs)
+        return ti.analysis.label.predict_transport_mode(self._obj, *args, **kwargs)
 
     def calculate_modal_split(self, *args, **kwargs):
         """
